spotify-playlist-creator/requirements.txt
numpy==1.21.0
pandas==1.3.0
requests==2.31.0
openai==1.3.0
flask==3.0.0
python-dotenv==1.0.0